# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Finance Management API",
        "description": "Backend API for the personal finance management app.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Base URL with port for local development"
        }
    ],
    "paths": {
        "/api/v1/transactions/": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "List transactions (paginated)",
                "description": "Paginated and filterable transaction list for the authenticated user.\n\nSupports search, category filter, amount range, date range, type (credit/debit),\nand sorting. The response shape is standardized for the mobile app.",
                "operationId": "list_transactions_api_v1_transactions__get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "description": "Page number (1-based)",
                            "default": 1,
                            "title": "Page"
                        },
                        "description": "Page number (1-based)"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1,
                            "description": "Items per page",
                            "default": 20,
                            "title": "Limit"
                        },
                        "description": "Items per page"
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Filter by category UUID",
                            "title": "Category"
                        },
                        "description": "Filter by category UUID"
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Search merchant name",
                            "title": "Search"
                        },
                        "description": "Search merchant name"
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Sort options: recent | highest",
                            "default": "recent",
                            "title": "Sort"
                        },
                        "description": "Sort options: recent | highest"
                    },
                    {
                        "name": "min_amount",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "number"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Minimum transaction amount",
                            "title": "Min Amount"
                        },
                        "description": "Minimum transaction amount"
                    },
                    {
                        "name": "max_amount",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "number"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Maximum transaction amount",
                            "title": "Max Amount"
                        },
                        "description": "Maximum transaction amount"
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "ISO start date",
                            "title": "Start Date"
                        },
                        "description": "ISO start date"
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "ISO end date",
                            "title": "End Date"
                        },
                        "description": "ISO end date"
                    },
                    {
                        "name": "month",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "maximum": 12,
                            "description": "Month number (1-12); requires `year` parameter",
                            "title": "Month"
                        },
                        "description": "Optional calendar month filter"
                    },
                    {
                        "name": "year",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 2000,
                            "description": "Year for month filter; must accompany `month`",
                            "title": "Year"
                        },
                        "description": "Optional calendar year filter"
                    },
                    {
                        "name": "tx_type",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "credit|debit",
                            "title": "Tx Type"
                        },
                        "description": "credit|debit"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Create a new transaction",
                "description": "Manual transaction creation. Amount positive -> credit, negative -> debit.",
                "operationId": "create_transaction_api_v1_transactions__post",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateTransactionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CreateTransactionResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/transactions/merchants": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Merchant autocomplete suggestions",
                "description": "Return distinct merchant names matching the query (case‑insensitive), up to 10 results.",
                "operationId": "merchant_suggestions_api_v1_transactions_merchants_get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "description": "Search query for merchant name",
                            "title": "Q"
                        },
                        "description": "Search query for merchant name"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/transactions/{transaction_id}": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Get a single transaction",
                "description": "Returns full detail for one transaction, including categorization.",
                "operationId": "get_transaction_api_v1_transactions__transaction_id__get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "transaction_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Transaction Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/transactions/{transaction_id}/suggestions": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Get Top-3 category suggestions",
                "description": "Returns Top-3 behavioral category suggestions for an UNCLASSIFIED transaction.\n\nIf suggestions are already stored in transaction_categorizations, returns those.\nOtherwise calls the behavior engine live (requires user to have ≥ 5 transactions\nof history — engine returns empty list if not).\n\nAlways returns requires_user_confirmation = true.\nOnly meaningful for UNCLASSIFIED transactions.",
                "operationId": "get_suggestions_api_v1_transactions__transaction_id__suggestions_get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "transaction_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Transaction Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreateTransactionRequest": {
                "properties": {
                    "merchant": {
                        "type": "string",
                        "title": "Merchant"
                    },
                    "amount": {
                        "type": "number",
                        "title": "Amount"
                    },
                    "category_id": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Category Id"
                    },
                    "date": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Date"
                    },
                    "notes": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Notes"
                    }
                },
                "type": "object",
                "required": [
                    "merchant",
                    "amount",
                    "category_id",
                    "date",
                    "notes"
                ],
                "title": "CreateTransactionRequest"
            },
            "CreateTransactionResponse": {
                "properties": {
                    "transaction_id": {
                        "type": "string",
                        "title": "Transaction Id"
                    },
                    "status": {
                        "type": "string",
                        "title": "Status"
                    }
                },
                "type": "object",
                "required": [
                    "transaction_id",
                    "status"
                ],
                "title": "CreateTransactionResponse"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}